PublicBaseURL hotfix (#117)
A hotfix for the PublicBaseURL parameter, where the update script runs an older version and inserts the placeholder value into a live config. Running the update function again, or waiting for it to run automatically, should remove the placeholder value and replace it with the correct value from the LME config.