7.16.2
Minor version bump to Elastic version 7.16.2 to resolve issues with vulnerability scanners flagging the service as vulnerable to the recent log4j vulnerability. This was addressed by Elastic in 7.16.1 through a series of mitigations, but as the version number hadn't changed this was still showing as vulnerable in some instances.